Hello
Hi, I'm trying to understand your project and wish to later fork and change the sign thing to a gui instead. Do you have any tips on where to start to try and fully understand so I can start coding/editing it myself? Or if you could possibly do it yourself that'd be great.